+++ src/tablecloth/api/fold_unroll.clj
@@ -32,5 +32,5 @@
    (let [colnames (column-names ds columns-selector)
          colnames-set (set colnames)]
      (if (grouped? ds)
-       (process-group-data ds #(process-unroll % colnames-set colnames options))
+       (process-group-data ds #(process-unroll % colnames-set colnames options) (:parallel options))
        (process-unroll ds colnames-set colnames options)))))